Fix Chip_RGU_TriggerReset M0 release bug
When the M0 APP/SUB system is still in reset and a reset was issued to
some other peripheral via Chip_RGU_TriggerReset a zero was written to
the specific bits which release the M0 cores from reset as a unwanted
side effect.

This is fixed by querying the reset status of the M0 cores and masking a
one at the right bits in the reset registers.